move findtiles from gcmfaces_nctiles.jl to gcmfaces_grids.jl
