+++ src/MeshArrays.jl
@@ -11,7 +11,7 @@ include("gcmfaces_exch.jl");
 include("gcmfaces_convert.jl");
 include("gcmfaces_IO.jl");
 include("gcmfaces_demo.jl");
-include("gcmfaces_nctiles.jl");
+
 
 export AbstractGcmfaces, gcmfaces, gcmsubset, fsize, fijind
 export exchange, gradient, convergence, smooth, mask
@@ -24,6 +24,7 @@ export read_bin, convert2array, convert2gcmfaces
 export exch_UV
 #The following codes add dependencies to Plots & NetCDF.
 #include("gcmfaces_plot.jl");
+#include("gcmfaces_nctiles.jl");
 
 
 end # module
